+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [1.2.2] - 2016-11-16
+- Added Maestro gateway.
+- Filter gateway description to show mandate notice also when description is empty.
+
 ## [1.2.1] - 2016-10-20
 - Added experimental support for WooCommerce Subscriptions / recurring payments.
 - Restore compatibility with WooCommerce versions < 2.2.0.
@@ -66,7 +70,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2014-12-19
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/woocommerce/compare/1.2.1...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/woocommerce/compare/1.2.2...HEAD
+[1.2.2]: https://github.com/wp-pay-extensions/woocommerce/compare/1.2.1...1.2.2
 [1.2.1]: https://github.com/wp-pay-extensions/woocommerce/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/wp-pay-extensions/woocommerce/compare/1.1.9...1.2.0
 [1.1.9]: https://github.com/wp-pay-extensions/woocommerce/compare/1.1.8...1.1.9